extract VideoFramePool from wgpu
