+++ src/window/mod.rs
@@ -1,11 +1,12 @@
-use std::sync::Arc;
+use std::{mem::size_of, ops::Deref, sync::Arc};
 
 use anyhow::Result;
+use thingbuf::{Recycle, ThingBuf};
 use winit::event_loop::EventLoop;
 
 use crate::input::keys::{KeyCode, Modifiers};
 
-use self::egui_winit_wgpu::{Renderer, VideoFramePool};
+use self::egui_winit_wgpu::Renderer;
 pub mod egui_winit_wgpu;
 mod winit_impl;
 
@@ -54,3 +55,45 @@ pub async fn create_state(
 
     Renderer::new(Arc::new(window_builder.build(event_loop)?), frame_pool).await
 }
+
+use crate::nes_state::VideoFrame;
+
+#[derive(Debug)]
+pub struct TFrameRecycle<const N: usize>;
+impl<const N: usize> Recycle<[u8; N]> for TFrameRecycle<N> {
+    fn new_element(&self) -> [u8; N] {
+        [0; N]
+    }
+
+    fn recycle(&self, _frame: &mut [u8; N]) {}
+}
+#[derive(Debug)]
+pub struct BytePool<const N: usize>(Arc<ThingBuf<[u8; N], TFrameRecycle<N>>>);
+
+impl<const N: usize> BytePool<N> {
+    pub fn new() -> Self {
+        Self(Arc::new(ThingBuf::with_recycle(2, TFrameRecycle)))
+    }
+}
+
+impl<const N: usize> Deref for BytePool<N> {
+    type Target = Arc<ThingBuf<[u8; N], TFrameRecycle<N>>>;
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl<const N: usize> Clone for BytePool<N> {
+    fn clone(&self) -> Self {
+        Self(Arc::clone(&self.0))
+    }
+}
+
+impl<const N: usize> Default for BytePool<N> {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+const VIDEO_FRAME_SIZE: usize = size_of::<VideoFrame>();
+pub type VideoFramePool = BytePool<VIDEO_FRAME_SIZE>;
